pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "claudette"
version = "1.3.0"
description = "Claude Code compatible CLI with preprocessing, multi-backend plugins, and cost analytics"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
authors = [
    {name = "Claude Flow Project", email = "contact@claudeflow.dev"},
]
keywords = ["claude", "ai", "cli", "preprocessing", "cost-analytics"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: System :: Systems Administration",
    "Topic :: Utilities",
]

dependencies = [
    "openai>=1.30.0",
    "pyyaml>=6.0",
    "tiktoken>=0.5.0",
    "click>=8.0.0",
    "pydantic>=2.0.0",
    "aiohttp>=3.8.0",
    "requests>=2.25.0",
    "rich>=13.0.0",
    "flask>=2.3.0",
    "plotly>=5.17.0",
    "pandas>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "pytest-benchmark>=4.0.0",
    "psutil>=5.9.0",
    "pylint>=2.15.0",
    "mypy>=1.0.0",
    "black>=23.12.0",
    "isort>=5.13.0",
    "flake8>=7.0.0",
    "bandit[toml]>=1.7.5",
    "pre-commit>=3.6.0",
    "commitizen>=3.13.0",
    "build>=0.10.0",
    "twine>=4.0.0",
    "mkdocs>=1.5.3",
    "mkdocs-material>=9.5.0",
    "mkdocstrings[python]>=0.24.0",
]

[project.urls]
Homepage = "https://github.com/ruvnet/claude-flow"
Repository = "https://github.com/ruvnet/claude-flow.git"
Documentation = "https://github.com/ruvnet/claude-flow/tree/main/docs"
"Bug Reports" = "https://github.com/ruvnet/claude-flow/issues"
Changelog = "https://github.com/ruvnet/claude-flow/blob/main/CHANGELOG.md"

[project.scripts]
claudette = "claudette.main:cli"
claude = "claudette.main:cli"

[tool.setuptools.packages.find]
include = ["claudette*"]
exclude = ["tests*", "docs*", "scripts*"]

[tool.setuptools.package-data]
claudette = ["*.yaml", "*.yml", "*.json"]

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --strict-markers"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    "integration: marks tests as integration tests",
    "benchmark: marks tests as benchmark tests",
    "slow: marks tests as slow running",
    "security: marks tests as security validation",
    "performance: marks tests as performance regression",
    "compatibility: marks tests as cross-platform compatibility",
    "resilience: marks tests as error handling and recovery",
]

[tool.pylint.messages_control]
disable = ["C0114", "C0115", "C0116"]  # Missing docstrings
min-similarity-lines = 10

[tool.pylint.format]
max-line-length = 100

[tool.pylint.design]
max-args = 8
max-locals = 20

[tool.black]
line-length = 100
target-version = ['py311', 'py312']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 100
known_first_party = ["claudette"]
known_third_party = ["openai", "pyyaml", "tiktoken", "click", "pydantic", "aiohttp", "requests", "rich", "flask", "plotly", "pandas"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true
show_error_codes = true
pretty = true

[[tool.mypy.overrides]]
module = ["tests.*"]
disallow_untyped_defs = false

[tool.coverage.run]
source = ["claudette"]
omit = ["*/tests/*", "*/test_*.py"]

[tool.coverage.report]
minimum = 85
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]

[tool.build]
exclude = [
    "tests/",
    "docs/",
    "scripts/",
    ".github/",
    "*.md",
    "Makefile",
]